test_requirements.txt
ipdb
fabric
mixer
pytest-cov
pytest-django
pytest
